requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
gspread
google-auth
pdfplumber
openai
streamlit
plotly
flask
